LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/implicit-submission.optional-expected.txt





FAIL Submit event with a submit button promise_test: Unhandled rejection with value: object "ReferenceError: Can't find variable: SubmitEvent"
FAIL Submit event with no submit button assert_equals: expected (object) null but got (undefined) undefined
